Added option to output SNPs in 012 format
I just added the option, then added some if statements in the function that prints the output.

I checked to make sure that the output is consistent (the genotypes) and that the output can be read by genoscapeRtools::read012().

Everything checks out.  Now we can use this to start simulating large SNP data sets that we can send to snppickr.